Infra RunEnv should be SDK env (#1502)
