+++ typescript/infra/scripts/utils.ts
@@ -7,7 +7,6 @@ import {
   ChainName,
   IChainConnection,
   MultiProvider,
-  coreEnvironments,
   objMap,
   promiseObjAll,
 } from '@hyperlane-xyz/sdk';
@@ -59,13 +58,6 @@ export function getCoreEnvironmentConfig<Env extends DeployEnvironment>(
 export async function getEnvironment() {
   return assertEnvironment(await getEnvironmentFromArgs());
 }
-
-type SdkEnvironment = keyof typeof coreEnvironments;
-export const deployEnvToSdkEnv: Record<DeployEnvironment, SdkEnvironment> = {
-  mainnet2: 'mainnet',
-  testnet3: 'testnet',
-  test: 'test',
-};
 
 export async function getEnvironmentConfig() {
   return getCoreEnvironmentConfig(await getEnvironment());
